+++ src/webapp/components/action-wizard/steps/GeneralInfoStep.tsx
@@ -3,7 +3,7 @@ import { TextField } from "@material-ui/core";
 import React, { useCallback, useEffect, useState } from "react";
 import styled from "styled-components";
 import { SyncAction } from "../../../../domain/entities/actions/SyncAction";
-import { DataMart } from "../../../../domain/entities/xmart/XMart";
+import { DataMart } from "../../../../domain/entities/xmart/DataMart";
 import i18n from "../../../../locales";
 import { Dictionary } from "../../../../types/utils";
 import { useAppContext } from "../../../contexts/app-context";
