+++ src/webapp/components/mapping-template-wizard/steps/GeneralInfoStep.tsx
@@ -2,7 +2,7 @@ import { useSnackbar } from "@eyeseetea/d2-ui-components";
 import { makeStyles, TextField } from "@material-ui/core";
 import React, { useCallback, useEffect, useState } from "react";
 import { SyncAction } from "../../../../domain/entities/actions/SyncAction";
-import { DataMart } from "../../../../domain/entities/xmart/XMart";
+import { DataMart } from "../../../../domain/entities/xmart/DataMart";
 import i18n from "../../../../locales";
 import { Dictionary } from "../../../../types/utils";
 import { useAppContext } from "../../../contexts/app-context";
